Remove MutexTestTrait
