+++ tests/MutexTest.php
@@ -5,12 +5,67 @@ declare(strict_types=1);
 namespace Yiisoft\Mutex\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Mutex\MutexInterface;
 use Yiisoft\Mutex\Tests\Mocks\Mutex;
 
 final class MutexTest extends TestCase
 {
-    use MutexTestTrait;
+    public function testMutexAcquire(): void
+    {
+        $mutex = $this->createMutex('testMutexAcquire');
+
+        $this->assertTrue($mutex->acquire());
+        $mutex->release();
+    }
+
+    public function testThatMutexLockIsWorking(): void
+    {
+        $mutexOne = $this->createMutex('testThatMutexLockIsWorking');
+        $mutexTwo = $this->createMutex('testThatMutexLockIsWorking');
+
+        $this->assertTrue($mutexOne->acquire());
+        $this->assertFalse($mutexTwo->acquire());
+        $mutexOne->release();
+        $mutexTwo->release();
+
+        $this->assertTrue($mutexTwo->acquire());
+        $mutexTwo->release();
+    }
+
+    public function testThatMutexLockIsWorkingOnTheSameComponent(): void
+    {
+        $mutex = $this->createMutex('testThatMutexLockIsWorkingOnTheSameComponent');
+
+        $this->assertTrue($mutex->acquire());
+        $this->assertFalse($mutex->acquire());
+
+        $mutex->release();
+        $mutex->release();
+    }
+
+    public function testTimeout(): void
+    {
+        $mutexOne = $this->createMutex(__METHOD__);
+        $mutexTwo = $this->createMutex(__METHOD__);
+
+        $this->assertTrue($mutexOne->acquire());
+        $microtime = microtime(true);
+        $this->assertFalse($mutexTwo->acquire(1));
+        $diff = microtime(true) - $microtime;
+        $this->assertTrue($diff >= 1 && $diff < 2);
+        $mutexOne->release();
+        $mutexTwo->release();
+    }
+
+    public function testDeleteLockFile(): void
+    {
+        $mutex = $this->createMutex('testDeleteLockFile');
+
+        $mutex->acquire();
+        $this->assertFileExists($mutex->getFile());
+
+        $mutex->release();
+        $this->assertFileDoesNotExist($mutex->getFile());
+    }
 
     public function testDestruct(): void
     {
@@ -25,7 +80,7 @@ final class MutexTest extends TestCase
         $this->assertFileDoesNotExist($file);
     }
 
-    protected function createMutex(string $name): MutexInterface
+    private function createMutex(string $name): Mutex
     {
         return new Mutex($name);
     }
